@@ -1,11 +1,11 @@
-let React = require('react');
+import React, { Component } from 'react';
 
-module.exports = React.createClass({
-  render: function() {
+export default class Footer extends Component {
+  render() {
     return (
       <footer className="footer">
         <p>A product of <a href="https://utensils.io">utensils.io</a> &copy;  2015</p>
       </footer>
     )
   }
-});
+};
