@@ -3,6 +3,6 @@ defmodule Hospital.PageControllerTest do
 
   test "GET /" do
     conn = get conn(), "/"
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
+    assert html_response(conn, 200) =~ ~s(<div id="root")
   end
 end
