@@ -8,11 +8,30 @@ defmodule Hospital.HealthCheckController do
   alias Hospital.User
 
   plug Guardian.Plug.EnsureAuthenticated, %{ on_failure: { SessionController, :unauthenticated_api } }
-  plug Guardian.Plug.EnsurePermissions, default: [:read, :write], on_failure: { SessionController, :forbidden_api }
   plug :retrieve_health_checks
 
   def index(conn, _params) do
     render(conn, "index.json", health_checks: conn.assigns[:health_checks])
+  end
+
+  def new(conn, _params) do
+    changeset = HealthCheck.create_changeset(%HealthCheck{})
+    render(conn, "new.html", changeset: changeset)
+  end
+
+  def create(conn, %{"health_check" => params}) do
+    user = Guardian.Plug.current_resource(conn)
+    params = Map.put(params, "user_id", user.id)
+    changeset = HealthCheck.create_changeset(%HealthCheck{}, params)
+
+    case Repo.insert(changeset) do
+      {:ok, _health_check} ->
+        conn
+        |> put_flash(:info, "Health check created successfully.")
+        |> redirect(to: health_check_path(conn, :index))
+      {:error, changeset} ->
+        render(conn, "new.html", changeset: changeset)
+    end
   end
 
   def show(conn, _) do
@@ -36,7 +55,7 @@ defmodule Hospital.HealthCheckController do
             where: h.user_id == ^user_id and h.id == ^id
 
     case Repo.one(query) do
-      {:ok, health_check} ->
+      %HealthCheck{} = health_check ->
         assign(conn, :health_check, health_check)
       nil ->
         SessionController.forbidden_api(conn)
