@@ -2,8 +2,9 @@ defmodule Hospital.Report do
   use Hospital.Web, :model
 
   schema "reports" do
+    field :checked_at, Ecto.DateTime
     field :successful, :boolean, default: false
-    field :results, :map
+    field :response_time, :float
 
     belongs_to :health_check, Hospital.HealthCheck
     belongs_to :medic, Hospital.Medic
@@ -11,8 +12,8 @@ defmodule Hospital.Report do
     timestamps
   end
 
-  @required_fields ~w(successful medic_id health_check_id)
-  @optional_fields ~w(results)
+  @required_fields ~w(successful response_time medic_id health_check_id)
+  @optional_fields ~w(checked_at)
 
   def create_changeset(model, params \\ :empty) do
     model
